Extract reusable dummy cast text members
