+++ Test/LingoEngine.Tests/CsvImporterTests.cs
@@ -12,6 +12,7 @@ using LingoEngine.Tools;
 using LingoEngine.Core;
 using LingoEngine.Primitives;
 using LingoEngine.Sprites;
+using LingoEngine.Tests.Casts;
 using Xunit;
 
 namespace LingoEngine.Tests;
@@ -95,137 +96,5 @@ public class CsvImporterTests
         public string? ReadTextFile(string fileName) => File.Exists(fileName) ? File.ReadAllText(fileName) : null;
 
         public byte[]? ReadBytes(string fileName) => File.Exists(fileName) ? File.ReadAllBytes(fileName) : null;
-    }
-
-    private class DummyCast : ILingoCast
-    {
-        public string Name => "Dummy";
-        public string FileName { get; set; } = string.Empty;
-        public int Number => 1;
-        public PreLoadModeType PreLoadMode { get; set; }
-        public bool IsInternal => true;
-        public CastMemberSelection? Selection { get; set; }
-        public ILingoMembersContainer Member { get; } = new DummyMembersContainer();
-        public ILingoMember? LastAddedMember { get; private set; }
-        public event Action<ILingoMember>? MemberAdded;
-        public event Action<ILingoMember>? MemberDeleted;
-        public event Action<ILingoMember>? MemberNameChanged;
-        public void Dispose() { }
-        public T? GetMember<T>(int number) where T : class, ILingoMember => null;
-        public T? GetMember<T>(string name) where T : class, ILingoMember => null;
-        public int FindEmpty() => 1;
-        public ILingoMember Add(LingoMemberType type, int numberInCast, string name, string fileName = "", APoint regPoint = default)
-        {
-            var member = new DummyTextMember(this) { Name = name };
-            member.SetFileName(fileName);
-            LastAddedMember = member;
-            MemberAdded?.Invoke(member);
-            return member;
-        }
-        public T Add<T>(int numberInCast, string name, Action<T>? configure = null) where T : ILingoMember => throw new NotImplementedException();
-        public IEnumerable<ILingoMember> GetAll() => Array.Empty<ILingoMember>();
-        public void SwapMembers(int slot1, int slot2) { }
-        public void Save() { }
-    }
-
-    private class DummyMembersContainer : ILingoMembersContainer
-    {
-        public ILingoMember? this[int number] => null;
-        public ILingoMember? this[string name] => null;
-        public event Action<ILingoMember>? MemberAdded;
-        public event Action<ILingoMember>? MemberDeleted;
-        public T? Member<T>(int number) where T : class, ILingoMember => null;
-        public T? Member<T>(string name) where T : class, ILingoMember => null;
-    }
-
-    private class DummyTextMember : ILingoMemberTextBase, ILingoMemberTextBaseInteral
-    {
-        private readonly ILingoCast _cast;
-        private readonly DummyFrameworkMember _framework = new();
-        private string _fileName = string.Empty;
-
-        public DummyTextMember(ILingoCast cast) => _cast = cast;
-
-        public ILingoFrameworkMember FrameworkObj => _framework;
-        public string Name { get; set; } = string.Empty;
-        public int Number => 0;
-        public DateTime CreationDate => DateTime.Now;
-        public DateTime ModifiedDate => DateTime.Now;
-        public bool Hilite => false;
-        public int CastLibNum => 0;
-        public int NumberInCast => 0;
-        public int PurgePriority { get; set; }
-        public APoint RegPoint { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
-        public long Size { get; set; }
-        public string Comments { get; set; } = string.Empty;
-        public string FileName => _fileName;
-        public LingoMemberType Type { get; set; } = LingoMemberType.Text;
-        public string CastName => string.Empty;
-        public ILingoCast Cast => _cast;
-        public void CopyToClipBoard() { }
-        public void Erase() { }
-        public void ImportFileInto() { }
-        public void Move() { }
-        public void PasteClipBoardInto() { }
-        public void Preload() { Preloaded = true; }
-        public void Unload() { }
-        public ILingoMember Duplicate(int? newNumber = null) => this;
-        public ILingoMember? GetMemberInCastByOffset(int numberOffset) => null;
-        public bool IsPixelTransparent(int x, int y) => false;
-        public void Dispose() { }
-        public IAbstTexture2D? TextureLingo => null;
-        public IAbstTexture2D? RenderToTexture(LingoInkType ink, AColor transparentColor) => null;
-        public string Text { get; set; } = string.Empty;
-        public void SetTextMD(AbstMarkdownData data)
-        {
-            Text = data.Markdown;
-            MarkdownData = data;
-        }
-        public LingoLines Line => new(() => { });
-        public LingoWords Word => new(Text);
-        public LingoChars Char => new();
-        public void Clear() { Text = string.Empty; }
-        public void Copy() { }
-        public void Cut() { }
-        public void Paste() { }
-        public void InsertText(string text) { }
-        public void ReplaceSelection(string replacement) { }
-        public void SetSelection(int start, int end) { }
-        public void RequireRedraw() { }
-        public void InitDefaults() { }
-        public IAbstTexture2D? GetTexture() => null;
-        public bool Editable { get; set; }
-        public bool WordWrap { get; set; }
-        public int ScrollTop { get; set; }
-        public string Font { get; set; } = string.Empty;
-        public int FontSize { get; set; }
-        public LingoTextStyle FontStyle { get; set; }
-        public AColor TextColor { get; set; }
-        public bool Bold { get; set; }
-        public bool Italic { get; set; }
-        public bool Underline { get; set; }
-        public AbstTextAlignment Alignment { get; set; }
-        public int Margin { get; set; }
-        public void ChangesHasBeenApplied() { }
-        public void LoadFile() { LoadFileCalled = true; }
-        public bool Preloaded { get; private set; }
-        public bool LoadFileCalled { get; private set; }
-        public AbstMarkdownData? MarkdownData { get; private set; }
-        public void SetFileName(string name) => _fileName = name;
-    }
-
-    private class DummyFrameworkMember : ILingoFrameworkMember
-    {
-        public bool IsLoaded => true;
-        public void CopyToClipboard() { }
-        public void Erase() { }
-        public void ImportFileInto() { }
-        public void PasteClipboardInto() { }
-        public void Preload() { }
-        public void ReleaseFromSprite(LingoSprite2D lingoSprite) { }
-        public void Unload() { }
-        public bool IsPixelTransparent(int x, int y) => false;
     }
 }
